@@ -11,29 +11,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.15",
-    "@angular/common": "~7.2.15",
-    "@angular/compiler": "~7.2.15",
-    "@angular/core": "~7.2.15",
-    "@angular/forms": "~7.2.15",
-    "@angular/platform-browser": "~7.2.15",
-    "@angular/platform-browser-dynamic": "~7.2.15",
-    "@angular/router": "~7.2.15",
+    "@angular/animations": "~8.2.4",
+    "@angular/common": "~8.2.4",
+    "@angular/compiler": "~8.2.4",
+    "@angular/core": "~8.2.4",
+    "@angular/forms": "~8.2.4",
+    "@angular/platform-browser": "~8.2.4",
+    "@angular/platform-browser-dynamic": "~8.2.4",
+    "@angular/router": "~8.2.4",
     "core-js": "^2.5.4",
     "rxjs": "~6.5.2",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.11.0",
-    "@angular-devkit/build-ng-packagr": "~0.11.0",
-    "@angular/cli": "~7.3.9",
-    "@angular/compiler-cli": "~7.2.15",
-    "@angular/language-service": "~7.2.15",
+    "@angular-devkit/build-angular": "~0.803.2",
+    "@angular-devkit/build-ng-packagr": "~0.803.2",
+    "@angular/cli": "~8.3.2",
+    "@angular/compiler-cli": "~8.2.4",
+    "@angular/language-service": "~8.2.4",
     "@types/node": "~8.9.4",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.1.1",
@@ -41,12 +41,11 @@
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^4.2.0",
+    "ng-packagr": "^5.4.0",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
-    "tsickle": ">=0.29.0",
-    "tslib": "^1.9.0",
+    "tsickle": "^0.37.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.5.3"
   }
 }
